refactor(trello): split into config/client/verifier modules (T-18)
Split the trello config.py monolith into the vendor-first layout:
- config.py — TrelloConfig, build_trello_config, trello_config_from_env
- client.py — _request_json, validate_trello_connection, create_trello_card
- verifier.py — TrelloValidationResult, validate_trello_config
- __init__.py re-exports the same public API (callers unchanged)

verifier.py calls the client submodule via attribute lookup (import
integrations.trello.client as client) to avoid a package/submodule import
cycle and keep monkeypatch semantics.

Repoint test patch targets (unit + e2e) to integrations.trello.client.* and
the validator-path MigrationCase to integrations/trello/verifier.py.
