+++ integrations/trello/__init__.py
@@ -1,14 +1,18 @@
 """Trello integration package."""
 
+from integrations.trello.client import (
+    create_trello_card,
+    validate_trello_connection,
+)
 from integrations.trello.config import (
     DEFAULT_TRELLO_BASE_URL,
     TrelloConfig,
-    TrelloValidationResult,
     build_trello_config,
-    create_trello_card,
     trello_config_from_env,
+)
+from integrations.trello.verifier import (
+    TrelloValidationResult,
     validate_trello_config,
-    validate_trello_connection,
 )
 
 __all__ = [
