+++ integrations/trello/config.py
@@ -1,14 +1,9 @@
-import logging
+"""Trello connection settings and config builders."""
+
 import os
-from dataclasses import dataclass
 from typing import Any
 
-import httpx
 from pydantic import BaseModel, Field, field_validator
-
-from integrations._validation_helpers import report_validation_failure
-
-logger = logging.getLogger(__name__)
 
 DEFAULT_TRELLO_BASE_URL = "https://api.trello.com/1"
 
@@ -34,14 +29,6 @@ class TrelloConfig(BaseModel):
         return self.base_url.rstrip("/")
 
 
-@dataclass(frozen=True)
-class TrelloValidationResult:
-    """Result of validating a Trello integration."""
-
-    ok: bool
-    detail: str
-
-
 def build_trello_config(raw: dict[str, Any] | None) -> TrelloConfig:
     """Build a normalized Trello config object from env/store data."""
     return TrelloConfig.model_validate(raw or {})
@@ -64,93 +51,3 @@ def trello_config_from_env() -> TrelloConfig | None:
             "list_id": os.getenv("TRELLO_LIST_ID", "").strip(),
         }
     )
-
-
-def _request_json(
-    config: TrelloConfig,
-    method: str,
-    path: str,
-    *,
-    params: list[tuple[str, str | int | float | bool | None]] | None = None,
-    json: dict[str, Any] | None = None,
-) -> Any:
-    request_params: list[tuple[str, str | int | float | bool | None]] = [
-        ("key", config.api_key),
-        ("token", config.token),
-    ]
-    if params:
-        request_params.extend(params)
-
-    url = f"{config.api_base_url}{path}"
-    response = httpx.request(
-        method,
-        url,
-        params=request_params,
-        json=json,
-        timeout=config.timeout_seconds,
-    )
-    response.raise_for_status()
-    return response.json()
-
-
-def validate_trello_connection(
-    *,
-    config: TrelloConfig,
-) -> dict[str, Any]:
-    """Validate Trello connection with a lightweight member query."""
-    payload = _request_json(config, "GET", "/members/me")
-    return payload if isinstance(payload, dict) else {}
-
-
-def validate_trello_config(config: TrelloConfig) -> TrelloValidationResult:
-    """Validate Trello connectivity."""
-    if not config.api_key:
-        return TrelloValidationResult(ok=False, detail="Trello API key is required.")
-    if not config.token:
-        return TrelloValidationResult(ok=False, detail="Trello token is required.")
-
-    try:
-        member = validate_trello_connection(config=config)
-        username = member.get("username", "unknown")
-        return TrelloValidationResult(
-            ok=True,
-            detail=f"Trello connectivity successful. Authenticated as @{username}",
-        )
-    except httpx.HTTPStatusError as err:
-        detail = err.response.text.strip() or str(err)
-        return TrelloValidationResult(ok=False, detail=f"Trello validation failed: {detail}")
-    except Exception as err:
-        report_validation_failure(
-            err,
-            logger=logger,
-            integration="trello",
-            method="validate_trello_config",
-        )
-        return TrelloValidationResult(ok=False, detail=f"Trello validation failed: {err}")
-
-
-def create_trello_card(
-    *,
-    config: TrelloConfig,
-    name: str,
-    desc: str,
-    list_id: str | None = None,
-) -> dict[str, Any]:
-    """Create a Trello card."""
-    target_list_id = (list_id or config.list_id).strip()
-    if not target_list_id:
-        raise ValueError("A list_id must be provided either via argument or config.")
-
-    payload = _request_json(
-        config,
-        "POST",
-        "/cards",
-        params=[
-            ("idList", target_list_id),
-        ],
-        json={
-            "name": name,
-            "desc": desc,
-        },
-    )
-    return payload if isinstance(payload, dict) else {}
